added more Testing functionality, emulating distribution manager
