+++ healing_test/build.gradle
@@ -6,11 +6,12 @@ jar {
 dependencies {
     compile 'org.springframework:spring-jms:4.2.1.RELEASE'
     compile 'com.google.guava:guava:11.0.2'
+    compile 'org.springframework.boot:spring-boot-starter-activemq'
+    compile 'org.springframework.boot:spring-boot-starter-test'
     compile 'org.apache.activemq:activemq-all:5.4.2'
 
-    testCompile project (':plugin/healing')
+    compile project (':plugin/healing')
+
     testCompile group: 'junit', name: 'junit', version: '4.11'
-    testCompile 'org.apache.activemq:activemq-all:5.4.2'
-    testCompile 'org.springframework.boot:spring-boot-starter-web:1.3.2.RELEASE'
-    testCompile 'org.springframework.boot:spring-boot-starter-test:1.3.2.RELEASE'
+
 }
