+++ healing_test/src/main/resources/application.properties
@@ -1,0 +1,2 @@
+spring.activemq.in-memory=true
+spring.activemq.pool.enabled=false
